+++ src/main.h
@@ -572,21 +572,6 @@ static const unsigned int REJECT_CONFLICT = 0x102;
 /** Transaction cannot be committed on my fork */
 static const unsigned int REJECT_WRONG_FORK = 0x103;
 
-struct COrphanTx
-{
-    CTransactionRef ptx;
-    NodeId fromPeer;
-    int64_t nEntryTime; // BU - Xtreme Thinblocks: used for aging orphans out of the cache
-    uint64_t nOrphanTxSize;
-};
-// BU: begin creating separate critical section for orphan cache and untangling from cs_main.
-extern CCriticalSection cs_orphancache;
-extern std::map<uint256, COrphanTx> mapOrphanTransactions GUARDED_BY(cs_orphancache);
-extern std::map<uint256, std::set<uint256> > mapOrphanTransactionsByPrev GUARDED_BY(cs_orphancache);
-
-void EraseOrphanTx(uint256 hash) EXCLUSIVE_LOCKS_REQUIRED(cs_orphancache);
-// BU: end
-
 CBlockIndex *FindMostWorkChain();
 
 // BU cleaning up at destuction time creates many global variable dependencies.  Instead clean up in a function called
